Many fixes

+++ api/routes/file.mjs
@@ -1,13 +1,11 @@
 import express from "express"
 const { Router, Request, Response } = express;
 const route = Router();
-import Entity, { sanitize } from "entitystorage"
-import { getTimestamp } from "../../../../tools/date.mjs"
-import {default as fileService, tokens} from "../../services/file.mjs"
+import { sanitize } from "entitystorage"
+import { default as fileService, tokens } from "../../services/file.mjs"
 import Archiver from 'archiver';
 import moment from "moment"
-import {service as userService} from "../../../../services/user.mjs"
-import {validateAccess} from "../../../../services/auth.mjs"
+import { validateAccess } from "../../../../services/auth.mjs"
 import File from "../../models/file.mjs";
 import Folder from "../../models/folder.mjs";
 import FileOrFolder from "../../models/fileorfolder.mjs";
@@ -22,105 +20,119 @@ export default (app) => {
   })
 
   route.get("/tag/:tag", function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.read"})) return;
+    if (!validateAccess(req, res, { permission: "file.read" })) return;
     let results = FileOrFolder.allByTag(sanitize(req.params.tag))
-                              .filter(f => f.hasAccess(res.locals.user, 'r'))
-                              .map(c => c.toObj(res.locals.user))
+      .filter(f => f.hasAccess(res.locals.user, 'r'))
+      .map(c => c.toObj(res.locals.user))
     res.json(results)
   })
 
   route.get("/path", function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.read"})) return;
+    if (!validateAccess(req, res, { permission: "file.read" })) return;
+    Folder.userRoot(res.locals.user) //Will create user root if missing
     res.json(Folder.root().toObj(res.locals.user))
   })
 
   route.get("/path/*", function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.read"})) return;
+    if (!validateAccess(req, res, { permission: "file.read" })) return;
     let path = decodeURI(req.path.substring(5))
+    Folder.userRoot(res.locals.user) //Will create user root if missing
     let folder = path.substring(1).split("/").reduce((parent, name) => {
       return parent?.getChildFolderNamed(name, res.locals.user) || null
     }, Folder.root())
-
     res.json(folder?.toObj(res.locals.user) || null)
   })
 
   route.get("/query", function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.read"})) return;
-    res.json(fileService.search(req.query.filter).results.map(r => r.toObj(res.locals.user)))
+    if (!validateAccess(req, res, { permission: "file.read" })) return;
+    res.json(fileService.search(req.query.filter).results
+      .filter(f => f.hasAccess(res.locals.user, 'r'))
+      .map(r => r.toObj(res.locals.user)))
   })
 
   route.post("/tag/:tag/upload", (req, res, next) => {
-    if(!validateAccess(req, res, {permission: "file.upload"})) return;
+    if (!validateAccess(req, res, { permission: "file.upload" })) return;
     let files = []
     for (let filedef in req.files) {
       let fileObj = Array.isArray(req.files[filedef]) ? req.files[filedef] : [req.files[filedef]]
       for (let f of fileObj) {
-        let file = new File({...f, tag: req.params.tag, owner: res.locals.user})
-        files.push({id: file._id, hash: file.hash})
+        let file = new File({ ...f, tag: req.params.tag, owner: res.locals.user })
+        files.push({ id: file._id, hash: file.hash })
       }
     }
     res.json(files)
   })
 
   route.post(["/drop", "/drop/shared"], (req, res, next) => {
-    if(!validateAccess(req, res, {permission: "file.upload"})) return;
+    if (!validateAccess(req, res, { permission: "file.upload" })) return;
     let expirationDate = new Date()
     expirationDate.setDate(expirationDate.getDate() + 30)
     let files = []
     for (let filedef in req.files) {
       let fileObj = Array.isArray(req.files[filedef]) ? req.files[filedef] : [req.files[filedef]]
       for (let f of fileObj) {
-        let file = new File({...f, tag: "drop", expire: expirationDate.toISOString(), owner: res.locals.user})
+        let file = new File({ ...f, tag: "drop", expire: expirationDate.toISOString(), owner: res.locals.user })
         file.acl = req.path.endsWith("/shared") ? "r:shared;w:shared" : "r:private;w:private"
-        files.push({id: file._id, hash: file.hash})
+        files.push({ id: file._id, hash: file.hash })
       }
     }
     res.json(files)
   })
 
   route.delete("/drop/all", (req, res, next) => {
-    if(!validateAccess(req, res, {permission: "file.edit"})) return;
-    FileOrFolder.allByTag("drop").forEach(f => f.delete())
-    res.json({success: true})
+    if (!validateAccess(req, res, { permission: "file.edit" })) return;
+    FileOrFolder.allByTag("drop")
+      .filter(f => f.hasAccess(res.locals.user, 'w'))
+      .forEach(f => f.delete())
+    res.json({ success: true })
   })
 
   route.post("/folder/:id/upload", (req, res, next) => {
-    if(!validateAccess(req, res, {permission: "file.upload"})) return;
+    if (!validateAccess(req, res, { permission: "file.upload" })) return;
     let folder = Folder.lookup(sanitize(req.params.id))
-    if(!folder) throw "Unknown folder"
+    if (!folder) throw "Unknown folder"
+    if (!folder.validateAccess(res, 'w')) return;
     let files = []
     for (let filedef in req.files) {
       let fileObj = Array.isArray(req.files[filedef]) ? req.files[filedef] : [req.files[filedef]]
       for (let f of fileObj) {
-        let file = new File({...f, tag: req.params.tag, owner: res.locals.user, folder})
-        files.push({id: file._id, hash: file.hash})
+        if (folder.hasChildNamed(f.name)) continue
+        let file = new File({
+          ...f,
+          tag: req.query.tag || null,
+          tags: req.query.tags?.split(",").map(t => t.trim()) || null,
+          owner: res.locals.user,
+          folder
+        })
+        files.push({ id: file._id, hash: file.hash })
       }
     }
     res.json(files)
   })
 
   route.post("/upload-basic", function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.upload"})) return;
-    if(!req.query.tags) throw "Tags in the query are mandatory"
-    if(!req.query.hash) throw "Must provide hash in query";
-    let f = {name: req.query.name||"file", size: parseInt(req.header("Content-Length")), hash: req.query.hash, mimetype: req.query.mime || "application/x-binary", data: req}
-    let file = new File({...f, tags: req.query.tags?.split(",").map(t => t.trim()||[]), owner: res.locals.user})
-    res.json({id: file._id, hash: file.hash, name: f.name})
+    if (!validateAccess(req, res, { permission: "file.upload" })) return;
+    if (!req.query.tags) throw "Tags in the query are mandatory"
+    if (!req.query.hash) throw "Must provide hash in query";
+    let f = { name: req.query.name || "file", size: parseInt(req.header("Content-Length")), hash: req.query.hash, mimetype: req.query.mime || "application/x-binary", data: req }
+    let file = new File({ ...f, tags: req.query.tags?.split(",").map(t => t.trim() || []), owner: res.locals.user })
+    res.json({ id: file._id, hash: file.hash, name: f.name })
   })
 
   route.post("/:id/folders", function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.edit"})) return;
-    if (!req.body.name) throw "name is mandatory"
-    if(!req.params.id || isNaN(req.params.id)) throw "Parent folder is mandatory"
+    if (!validateAccess(req, res, { permission: "file.edit" })) return;
+    if (!sanitize(req.body.name)) throw "name is mandatory"
+    if (!req.params.id || isNaN(req.params.id)) throw "Parent folder is mandatory"
     let parent = Folder.lookup(sanitize(req.params.id))
-    if(!parent) throw "Parent doesn't exist";
-    if(parent.getChildFolderNamed(req.body.name)) throw "A file or folder with that name already exists"
+    if (!parent) throw "Parent doesn't exist";
+    if (!parent.validateAccess(res, 'w')) return;
+    if (parent.hasChildNamed(sanitize(req.body.name))) throw "A file or folder with that name already exists"
     let child = new Folder(req.body.name, res.locals.user, parent)
     res.json(child.toObj(res.locals.user))
   })
 
   route.get(['/dl/:id', '/dl/:id/:filename', '/download/:id', '/download/:id/:filename'], function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.read"})) return;
+    if (!validateAccess(req, res, { permission: "file.read" })) return;
     let file = File.lookupAccessible(sanitize(req.params.id), res.locals.user)
     if (!file) throw "Unknown file";
 
@@ -132,7 +144,7 @@ export default (app) => {
   });
 
   route.get(['/raw/:id', '/raw/:id/:filename'], function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.read"})) return;
+    if (!validateAccess(req, res, { permission: "file.read" })) return;
     let file = File.lookupAccessible(sanitize(req.params.id), res.locals.user)
     if (!file) throw "Unknown file";
 
@@ -144,8 +156,9 @@ export default (app) => {
   });
 
   route.get('/query/dl', function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.read"})) return;
-    let files = fileService.search(req.query.filter).results.filter(f => !f.tags.includes("folder"))
+    if (!validateAccess(req, res, { permission: "file.read" })) return;
+    let files = fileService.search(req.query.filter).results
+                           .filter(f => !f.tags.includes("folder") && f.hasAccess(res.locals.user, 'r'))
     if (files.length < 1) throw "No files in filter";
     let zip = Archiver('zip');
     for (let file of files) {
@@ -160,28 +173,30 @@ export default (app) => {
   });
 
   route.delete('/query', function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.edit"})) return;
+    if (!validateAccess(req, res, { permission: "file.edit" })) return;
     let files = fileService.search(req.query.filter)
     if (files.results.length < 1) return res.json(true);
 
     for (let file of files.results) {
+      if(!file.hasAccess(res.locals.user, 'w')) continue;
       file.delete()
     }
     res.json(true);
   });
 
   route.patch('/:id', function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.edit"})) return;
+    if (!validateAccess(req, res, { permission: "file.edit" })) return;
     let file = FileOrFolder.lookup(sanitize(req.params.id))
-    if(!file) throw "Unknown file or folder"
+    if (!file) throw "Unknown file or folder"
+    if(!file.validateAccess(res, 'w')) return;
 
-    if(req.body.name !== undefined && req.body.name) file.name = req.body.name
-    if(req.body.filename !== undefined && req.body.filename) file.name = req.body.filename
+    if (req.body.name !== undefined && req.body.name) file.name = req.body.name
+    if (req.body.filename !== undefined && req.body.filename) file.name = req.body.filename
 
     let tagList = !req.body.tags ? null : Array.isArray(req.body.tags) ? req.body.tags : typeof req.body.tags === "string" ? req.body.tags.split(",").map(t => t.trim()) : null;
-    if(tagList){
+    if (tagList) {
       tagList = tagList.map(t => `user-${t}`)
-      for(let t of file.tags.filter(t => t.startsWith("user-") && !tagList.includes(t))) 
+      for (let t of file.tags.filter(t => t.startsWith("user-") && !tagList.includes(t)))
         file.removeTag(t)
       file.tag(tagList)
     }
@@ -189,22 +204,24 @@ export default (app) => {
   })
 
   route.post('/:id/tags', function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.edit"})) return;
+    if (!validateAccess(req, res, { permission: "file.edit" })) return;
     let file = FileOrFolder.lookup(sanitize(req.params.id))
-    if(!file) throw "Unknown file"
+    if (!file) throw "Unknown file"
+    if(!file.validateAccess(res, 'w')) return;
     let tags = req.body.tag ? [req.body.tag] : (req.body.tags && Array.isArray(req.body.tags)) ? req.body.tags : []
-    for(let tag of tags){
-      if(typeof tag !== "string") continue;
+    for (let tag of tags) {
+      if (typeof tag !== "string") continue;
       file.tag(`user-${tag}`)
     }
     res.json(true)
   })
 
   route.delete('/:id/tags/:tag', function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.edit"})) return;
+    if (!validateAccess(req, res, { permission: "file.edit" })) return;
     let file = FileOrFolder.lookup(sanitize(req.params.id))
-    if(!file) throw "Unknown file"
-    if(!req.params.tag) throw "No tag provided"
+    if (!file) throw "Unknown file"
+    if (!req.params.tag) throw "No tag provided"
+    if(!file.validateAccess(res, 'w')) return;
 
     file.removeTag(`user-${req.params.tag}`)
 
@@ -212,16 +229,17 @@ export default (app) => {
   })
 
   route.delete('/:id', function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.edit"})) return;
+    if (!validateAccess(req, res, { permission: "file.edit" })) return;
     let file = FileOrFolder.lookup(sanitize(req.params.id))
-    if(!file) throw "Unknown file"
+    if (!file) throw "Unknown file"
+    if(!file.validateAccess(res, 'w')) return;
     file.delete();
 
     res.json(true)
   })
 
   route.get('/:id', async function (req, res, next) {
-    if(!validateAccess(req, res, {permission: "file.read"})) return;
+    if (!validateAccess(req, res, { permission: "file.read" })) return;
     let file = FileOrFolder.lookupAccessible(sanitize(req.params.id), res.locals.user)
     if (!file) return res.sendStatus(404);
     res.json(file.toObj(res.locals.user));
